Reorganized file structure
